requirements.txt
torch>=2.2.0,<2.4
transformers>=4.46.0,<5.0.0
scikit-learn>=1.8.0
PyYAML>=6.0
pandas>=2.2.0,<3.0.0
matplotlib>=3.10.0
plotly>=6.0.0
psutil>=5.9.0
numpy>=1.26.0,<2.0.0
safetensors>=0.5.0
